yeah

+++ scripts/deps.sh
@@ -8,11 +8,13 @@ XACT_EXT_DIR=${XACT_ROOT_DIR}/external
 function cm-build () {
     local dir_name="$1"
     if [[ "${dir_name}" == "" ]]; then
-        echo "specify dir name" &>2
+        echo "specify dir name" >&2
         return 1
     else
         pushd ${dir_name}
-        if [[ ! -d "build" ]]; then
+        if [[ -d "build" ]]; then
+            echo "${dir_name}/build exists; not rebuilding" >&2
+        else
             mkdir build
             pushd build
             cmake ../ && make -j8
@@ -25,12 +27,14 @@ function cm-build () {
 function cm-clean() {
     local dir_name="$1"
     if [[ "${dir_name}" == "" ]]; then
-        echo "specify dir name" &>2
+        echo "specify dir name" >&2
         return 1
     else
-        pushd ${dir_name}
-        rm -rf build
-        popd
+        if [[ -d ${dir_name} ]]; then
+            pushd ${dir_name}
+            rm -rf build
+            popd
+        fi
     fi
 }
 
@@ -40,6 +44,14 @@ CM_LIBS="googletest \
 
 
 function build-all() {
+    if [[ -f ${XACT_EXT_DIR}/googletest/CMakeLists.txt ]]; then
+        echo "git submodules are ok..." >&2
+    else
+        echo "updating submodules" >&2
+        pushd ${XACT_ROOT_DIR}
+        git submodule init && git submodule update
+        popd
+    fi
     for lib in ${CM_LIBS}; do
         cm-build ${XACT_EXT_DIR}/${lib}
     done
